refactor: extract new `<WorkspaceSelect>` component
Extract from `<ActiveWorkspaceSelect>` to allow for usage when selecting source workspaces.
